Implement and integrate new separate() (with optional remove) and cleanups along the way.
- Facade_Cluster.h and clustering_examine_x_boundary.cxx
  Return grouping but do not apply it in "examine".  Caller applies it.

- Rename Examine_overclustering -> Separate_overclustering to reflect the fact it mutates.

- Facade_Cluster.cxx remove old separate() entirely.  NaryTree::FacadeParent provides it.

- clustering_connect.cxx add assert to preempt the segfault in get_two_extreme_points() when cluster has none.

- add assert() in calling code on the cluster passed to separate to emphasize the fact it is removed.

- huge amount of general warnings fixes: int/size_t type warnings, hide (give up on) parenthesis, comment unused vars and static functions.

- Make some local functions static.

+++ clus/inc/WireCellClus/Facade_Util.h
@@ -100,7 +100,7 @@ namespace WireCell::PointCloud::Facade {
         // Sample points from first cloud at regular intervals
         int stride = std::max(1, (int)(one.points()[0].size() / 20)); // Sample ~20 points
         
-        for(int i = 0; i < one.points()[0].size(); i += stride) {
+        for(size_t i = 0; i < one.points()[0].size(); i += stride) {
             // Get K nearest neighbors from second cloud
             auto p1 = one.point(i);
             auto knn = two.kd().knn(5, p1); // Get 5 nearest neighbors
